@@ -3,6 +3,8 @@ require 'uri'
 
 describe CropsController, :type => :controller do
   it 'Should direct to a new page' do
+    user = FactoryGirl.create(:user)
+    sign_in user
     get 'new'
     expect(response).to render_template(:new)
     expect(response.status).to eq(200)
@@ -10,7 +12,7 @@ describe CropsController, :type => :controller do
 
   it 'Should redirect to crop_searches' do
     get 'index'
-    response.should redirect_to controller: 'crop_searches', action: 'search'
+    expect(response).to redirect_to controller: 'crop_searches', action: 'search'
     expect(response.status).to eq(302)
   end
 
@@ -23,33 +25,31 @@ describe CropsController, :type => :controller do
 
   it 'Should direct to create guide page after successful crop creation' do
     crop = FactoryGirl.attributes_for(:crop)
+    user = FactoryGirl.create(:user)
+    sign_in user
     post 'create', crop: crop
     expect(response.status).to eq(302)
-    response.should redirect_to "/en/guides/new?crop_id=#{assigns(:crop).id}"
+    expect(response).to redirect_to(
+      "/en/guides/new?crop_id=#{assigns(:crop).id}")
   end
 
   it 'Should redirect back to form after unsuccessful crop creation' do
     crop = FactoryGirl.attributes_for(:crop)
+    user = FactoryGirl.create(:user)
+    sign_in user
     crop[:name] = ""
     post 'create', crop: crop
     expect(response).to render_template(:new)
     expect(response.status).to eq(200)
   end
 
-  it 'should render an edit page if the user is admin' do
-    user = FactoryGirl.create(:user, admin: true)
+  it 'should render an edit page if the user is logged in' do
+    user = FactoryGirl.create(:user)
     sign_in user
     crop = FactoryGirl.create(:crop)
     get 'edit', id: crop.id
     expect(response).to render_template(:edit)
     expect(response.status).to eq(200)
-  end
-
-  it 'should not render an edit page if the user is not admin' do
-    crop = FactoryGirl.create(:crop)
-    get 'edit', id: crop.id
-    response.should redirect_to root_path
-    expect(response.status).to eq(302)
   end
 
   it 'should rerender the edit page if not all params are good' do
