@@ -6,9 +6,25 @@ class CropPolicy < ApplicationPolicy
     @crop = crop
   end
 
+  def new?
+    if @current_user
+      true
+    else
+      false
+    end
+  end
+
+  def create?
+    if @current_user
+      true
+    else
+      false
+    end
+  end
+
   def update?
     if @current_user
-      @current_user.admin
+      true
     else
       false
     end
@@ -20,7 +36,7 @@ class CropPolicy < ApplicationPolicy
   # To be
   def edit?
     if @current_user
-      @current_user.admin
+      true
     else
       false
     end
