@@ -4,14 +4,14 @@ module IntegrationHelper
       expect(page).to have_content(text)
     end
 
-    # def wait_for_ajax
-    #   Timeout.timeout(Capybara.default_wait_time) do
-    #     active = page.evaluate_script('angular.element.active')
-    #     until active == 0
-    #       active = page.evaluate_script('angular.element.active')
-    #     end
-    #   end
-    # end
+    def wait_for_ajax
+      Timeout.timeout(Capybara.default_wait_time) do
+        active = page.evaluate_script('angular.element.active')
+        until active == 0
+          active = page.evaluate_script('angular.element.active')
+        end
+      end
+    end
   end
 
   def self.included(receiver)
